Create migration script and replace overwrites with new methods
- replace update/delete_all overwrites with a new method specifically
  for doing this process with patient_teams sync
- Update tests to ensure syncing works properly
- Replace relevant update/delete_all with synced versions
- Create migration script for doing the initial patient_teams table
  population

+++ app/controllers/api/testing/teams_controller.rb
@@ -17,9 +17,9 @@ class API::Testing::TeamsController < API::Testing::BaseController
     log_destroy(ImmunisationImport.where(team:))
     log_destroy(ClassImport.where(team:))
 
-    log_destroy(SchoolMove.where(team:))
+    log_destroy(SchoolMove.where(team:), true)
     log_destroy(Consent.where(team:))
-    log_destroy(ArchiveReason.where(team:))
+    log_destroy(ArchiveReason.where(team:), true)
     log_destroy(Triage.where(team:))
 
     log_destroy(
@@ -29,16 +29,20 @@ class API::Testing::TeamsController < API::Testing::BaseController
 
     log_destroy(ConsentNotification.joins(:session).where(session: { team: }))
     log_destroy(SessionNotification.joins(:session).where(session: { team: }))
-    log_destroy(VaccinationRecord.joins(:session).where(session: { team: }))
+    log_destroy(
+      VaccinationRecord.joins(:session).where(sessions: { team: }),
+      true
+    )
 
     patient_ids = team.patients.pluck(:id)
 
     log_destroy(
-      PatientLocation.joins(location: :subteam).where(subteam: { team: })
+      PatientLocation.joins(location: :subteam).where(subteam: { team: }),
+      true
     )
 
     log_destroy(AccessLogEntry.where(patient_id: patient_ids))
-    log_destroy(ArchiveReason.where(patient_id: patient_ids))
+    log_destroy(ArchiveReason.where(patient_id: patient_ids), true)
     log_destroy(AttendanceRecord.where(patient_id: patient_ids))
     log_destroy(ConsentNotification.where(patient_id: patient_ids))
     log_destroy(GillickAssessment.where(patient_id: patient_ids))
@@ -47,13 +51,13 @@ class API::Testing::TeamsController < API::Testing::BaseController
     log_destroy(NotifyLogEntry.where(patient_id: nil))
     log_destroy(NotifyLogEntry.where(patient_id: patient_ids))
     log_destroy(PatientChangeset.where(patient_id: patient_ids))
-    log_destroy(PatientLocation.where(patient_id: patient_ids))
+    log_destroy(PatientLocation.where(patient_id: patient_ids), true)
     log_destroy(PatientSpecificDirection.where(patient_id: patient_ids))
     log_destroy(PDSSearchResult.where(patient_id: patient_ids))
     log_destroy(PreScreening.where(patient_id: patient_ids))
-    log_destroy(SchoolMove.where(patient_id: patient_ids))
+    log_destroy(SchoolMove.where(patient_id: patient_ids), true)
     log_destroy(SchoolMoveLogEntry.where(patient_id: patient_ids))
-    log_destroy(VaccinationRecord.where(patient_id: patient_ids))
+    log_destroy(VaccinationRecord.where(patient_id: patient_ids), true)
 
     log_destroy(SessionDate.joins(:session).where(session: { team: }))
 
@@ -61,11 +65,12 @@ class API::Testing::TeamsController < API::Testing::BaseController
     log_destroy(Patient.where(id: patient_ids))
     log_destroy(Parent.where.missing(:parent_relationships))
 
-    log_destroy(VaccinationRecord.joins(:batch).where(batch: { team: }))
+    log_destroy(VaccinationRecord.joins(:batch).where(batch: { team: }), true)
     log_destroy(Batch.where(team:))
 
     log_destroy(
-      VaccinationRecord.where(performed_ods_code: team.organisation.ods_code)
+      VaccinationRecord.where(performed_ods_code: team.organisation.ods_code),
+      true
     )
 
     unless keep_itself
@@ -77,7 +82,7 @@ class API::Testing::TeamsController < API::Testing::BaseController
       Location
         .joins(:subteam)
         .where(subteam: { team: })
-        .update_all(subteam_id: nil)
+        .update_all_with_patient_team_sync(subteam_id: nil)
 
       log_destroy(Subteam.where(team:))
 
@@ -93,10 +98,14 @@ class API::Testing::TeamsController < API::Testing::BaseController
 
   private
 
-  def log_destroy(query)
+  def log_destroy(query, with_patient_team_sync: false)
     where_clause = query.where_clause
     @log_time ||= Time.zone.now
-    query.delete_all
+    if with_patient_team_sync
+      query.delete_all_with_patient_team_sync
+    else
+      query.delete_all
+    end
     response.stream.write(
       "#{query.model.name}.where(#{where_clause.to_h}): #{Time.zone.now - @log_time}s\n"
     )
